# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Network Optimization Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface WarehouseCandidate {
  id: string;
  city: string;
  state: string;
  lat: number;
  lng: number;
}

export interface Customer {
  id: string;
  lat: number;
  lng: number;
  demand: number;
}

export interface Dataset {
  warehouses: WarehouseCandidate[];
  customers: Customer[];
}

export type WarehouseStatusEntryStatus = typeof WarehouseStatusEntryStatus[keyof typeof WarehouseStatusEntryStatus];


export const WarehouseStatusEntryStatus = {
  potential: 'potential',
  forced_open: 'forced_open',
  inactive: 'inactive',
} as const;

export interface WarehouseStatusEntry {
  warehouseId: string;
  status: WarehouseStatusEntryStatus;
}

export interface BandCoverage {
  band: number;
  percent: number;
}

export interface WarehouseUtilization {
  warehouseId: string;
  city: string;
  utilization: number;
}

export interface Assignment {
  customerId: string;
  warehouseId: string;
  distanceMi: number;
  band: number;
}

export type SolveResultStatus = typeof SolveResultStatus[keyof typeof SolveResultStatus];


export const SolveResultStatus = {
  optimal: 'optimal',
  infeasible: 'infeasible',
  solving: 'solving',
  error: 'error',
} as const;

export interface SolveResult {
  status: SolveResultStatus;
  openWarehouseIds: string[];
  assignments: Assignment[];
  objective: number;
  weightedAvgDistanceMi: number;
  bandCoverage: BandCoverage[];
  utilization: WarehouseUtilization[];
  runTimeSec: number;
  solverUsed: string;
  /** @nullable */
  infeasibilityReason: string | null;
}

export type ScenarioProblemType = typeof ScenarioProblemType[keyof typeof ScenarioProblemType];


export const ScenarioProblemType = {
  p_median: 'p_median',
  capacitated_flp: 'capacitated_flp',
  max_coverage: 'max_coverage',
  p_center: 'p_center',
  set_cover: 'set_cover',
} as const;

export type ScenarioSolver = typeof ScenarioSolver[keyof typeof ScenarioSolver];


export const ScenarioSolver = {
  cbc: 'cbc',
  highs: 'highs',
  glpk: 'glpk',
  gurobi: 'gurobi',
  scip: 'scip',
} as const;

export type ScenarioCapacityMode = typeof ScenarioCapacityMode[keyof typeof ScenarioCapacityMode];


export const ScenarioCapacityMode = {
  uniform: 'uniform',
  per_wh: 'per_wh',
} as const;

export interface Scenario {
  id: number;
  name: string;
  problemType: ScenarioProblemType;
  pValue: number;
  distanceBands: number[];
  solver: ScenarioSolver;
  gap: number;
  timeLimitSec: number;
  capacityMode: ScenarioCapacityMode;
  /** @nullable */
  uniformCapacity: number | null;
  warehouseStatuses: WarehouseStatusEntry[];
  result: SolveResult | null;
  createdAt: string;
  updatedAt: string;
}

export type ScenarioInputProblemType = typeof ScenarioInputProblemType[keyof typeof ScenarioInputProblemType];


export const ScenarioInputProblemType = {
  p_median: 'p_median',
  capacitated_flp: 'capacitated_flp',
  max_coverage: 'max_coverage',
  p_center: 'p_center',
  set_cover: 'set_cover',
} as const;

export type ScenarioInputSolver = typeof ScenarioInputSolver[keyof typeof ScenarioInputSolver];


export const ScenarioInputSolver = {
  cbc: 'cbc',
  highs: 'highs',
  glpk: 'glpk',
  gurobi: 'gurobi',
  scip: 'scip',
} as const;

export type ScenarioInputCapacityMode = typeof ScenarioInputCapacityMode[keyof typeof ScenarioInputCapacityMode];


export const ScenarioInputCapacityMode = {
  uniform: 'uniform',
  per_wh: 'per_wh',
} as const;

export interface ScenarioInput {
  name: string;
  problemType: ScenarioInputProblemType;
  /**
     * @minimum 1
     * @maximum 50
     */
  pValue: number;
  /** @minItems 1 */
  distanceBands: number[];
  solver: ScenarioInputSolver;
  /** @minimum 0 */
  gap: number;
  /** @minimum 1 */
  timeLimitSec: number;
  capacityMode: ScenarioInputCapacityMode;
  /** @nullable */
  uniformCapacity: number | null;
  warehouseStatuses: WarehouseStatusEntry[];
}

export type ScenarioUpdateProblemType = typeof ScenarioUpdateProblemType[keyof typeof ScenarioUpdateProblemType];


export const ScenarioUpdateProblemType = {
  p_median: 'p_median',
  capacitated_flp: 'capacitated_flp',
  max_coverage: 'max_coverage',
  p_center: 'p_center',
  set_cover: 'set_cover',
} as const;

export type ScenarioUpdateSolver = typeof ScenarioUpdateSolver[keyof typeof ScenarioUpdateSolver];


export const ScenarioUpdateSolver = {
  cbc: 'cbc',
  highs: 'highs',
  glpk: 'glpk',
  gurobi: 'gurobi',
  scip: 'scip',
} as const;

export type ScenarioUpdateCapacityMode = typeof ScenarioUpdateCapacityMode[keyof typeof ScenarioUpdateCapacityMode];


export const ScenarioUpdateCapacityMode = {
  uniform: 'uniform',
  per_wh: 'per_wh',
} as const;

export interface ScenarioUpdate {
  name?: string;
  problemType?: ScenarioUpdateProblemType;
  /**
     * @minimum 1
     * @maximum 50
     */
  pValue?: number;
  distanceBands?: number[];
  solver?: ScenarioUpdateSolver;
  gap?: number;
  timeLimitSec?: number;
  capacityMode?: ScenarioUpdateCapacityMode;
  /** @nullable */
  uniformCapacity?: number | null;
  warehouseStatuses?: WarehouseStatusEntry[];
}

export interface CompareInput {
  /** @minItems 2 */
  scenarioIds: number[];
}

export interface ScenarioMetrics {
  scenarioId: number;
  name: string;
  openSites: string[];
  weightedAvgDistanceMi: number;
  objective: number;
  bandDemandPercent: BandCoverage[];
  avgUtilization: number;
  solverStatus: string;
}

export interface CompareResult {
  scenarios: ScenarioMetrics[];
}