# lib/api-zod/src/generated/types/compareInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Network Optimization Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface CompareInput {
  /** @minItems 2 */
  scenarioIds: number[];
}