# lib/api-zod/src/generated/types/compareResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Network Optimization Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { ScenarioMetrics } from './scenarioMetrics';

export interface CompareResult {
  scenarios: ScenarioMetrics[];
}