# lib/api-zod/src/generated/types/scenarioMetrics.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Network Optimization Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { BandCoverage } from './bandCoverage';

export interface ScenarioMetrics {
  scenarioId: number;
  name: string;
  openSites: string[];
  weightedAvgDistanceMi: number;
  objective: number;
  bandDemandPercent: BandCoverage[];
  avgUtilization: number;
  solverStatus: string;
}