# lib/api-zod/src/generated/types/solveResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Network Optimization Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { Assignment } from './assignment';
import type { BandCoverage } from './bandCoverage';
import type { SolveResultStatus } from './solveResultStatus';
import type { WarehouseUtilization } from './warehouseUtilization';

export interface SolveResult {
  status: SolveResultStatus;
  openWarehouseIds: string[];
  assignments: Assignment[];
  objective: number;
  weightedAvgDistanceMi: number;
  bandCoverage: BandCoverage[];
  utilization: WarehouseUtilization[];
  runTimeSec: number;
  solverUsed: string;
  /** @nullable */
  infeasibilityReason: string | null;
}